requirements.txt
# Web Framework
fastapi
uvicorn[standard]
pydantic>=2.0.0
pydantic-settings

# HTTP Client
httpx
aiohttp

# Database
sqlalchemy>=2.0.0
alembic
asyncpg  # PostgreSQL async driver
aiosqlite  # SQLite async driver

# Crypto & Security
cryptography>=3.4.0
python-jose[cryptography]

# WebSocket
websockets

# Logging & Monitoring
loguru
prometheus-client

# Configuration
python-dotenv
pyyaml

# Testing
pytest
pytest-asyncio

# Development
black
isort
flake8
mypy

# Exchange APIs (CCXT for unified API access)
ccxt
